+++ packages/botkit/src/pages.tsx
@@ -14,6 +14,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 /** @jsxImportSource hono/jsx */
+import "./temporal.ts";
 import type { Context } from "@fedify/fedify/federation";
 import {
   type Announce,