+++ packages/botkit/tsdown.config.ts
@@ -11,10 +11,4 @@ export default defineConfig({
   format: "esm",
   platform: "node",
   unbundle: true,
-  outputOptions: {
-    intro: `
-      import { Temporal, toTemporalInstant } from "@js-temporal/polyfill";
-      Date.prototype.toTemporalInstant = toTemporalInstant;
-    `,
-  },
 });